convert to use pyrite (#49)
* annotate namepspace after apply

* bump version

* convert to use pyrite

+++ src/psykube/manifest/healthcheck/http.cr
@@ -1,7 +1,7 @@
 class Psykube::Manifest::Healthcheck::Http
   Manifest.mapping({
     path:    {type: String, default: "/"},
-    port:    {type: String | UInt16, default: "default"},
+    port:    {type: String | Int32, default: "default"},
     host:    String?,
     scheme:  String?,
     headers: {type: Hash(String, String), nilable: true},
