+++ src/psykube/manifest/healthcheck/tcp.cr
@@ -1,5 +1,5 @@
 class Psykube::Manifest::Healthcheck::Tcp
   Manifest.mapping({
-    port: {type: String | UInt16, default: "default"},
+    port: {type: String | Int32, default: "default"},
   })
 end
